дз по 5 раздделу
